Added ability to take back picture on other cards

+++ cardkeeper/cardkeeper/ViewModels/AddCardViewModel.cs
@@ -48,6 +48,24 @@ namespace cardkeeper.ViewModels
                 OnPropertyChanged("DisplayFrontImage");
             }
         }
+        private byte[] backImage;
+        private ImageSource displayBackImage;
+        public ImageSource DisplayBackImage
+        {
+            get
+            {
+                if (displayBackImage != null)
+                    return displayBackImage;
+                else
+                    return null;
+            }
+            set
+            {
+                if (displayBackImage != value)
+                    displayBackImage = value;
+                OnPropertyChanged("DisplayBackImage");
+            }
+        }
         public Card Card
         {
             get { return (Card)GetValue(CardProperty); }
@@ -58,6 +76,8 @@ namespace cardkeeper.ViewModels
 
         public ICommand SubmitButtonCommand { get; set; }
         public ICommand TakeFrontCardPhoto { get; set; }
+        public ICommand TakeBackCardPhoto { get; set; }
+
 
         public INavigation Navigation { get; set; }
 
@@ -67,8 +87,11 @@ namespace cardkeeper.ViewModels
             this.cardType = cardType;
             Card.Type = cardType;
             DisplayFrontImage = Card.DisplayFrontImage;
+            DisplayBackImage = Card.DisplayBackImage;
             SubmitButtonCommand = new Command(ValidateCard);
             TakeFrontCardPhoto = new Command(TakeFrontPhoto);
+            TakeBackCardPhoto = new Command(TakeBackPhoto);
+
         }
         public async void ValidateCard()
         {
@@ -91,7 +114,6 @@ namespace cardkeeper.ViewModels
                     }
                 case "Loyalty":
                 case "Membership":
-                case "Other":
                     {
                         if(Card.AccountNumber == null || Card.AccountNumber == "")
                         {
@@ -101,6 +123,23 @@ namespace cardkeeper.ViewModels
                         {
                             if (frontImage != null)
                                 Card.FrontImage = frontImage;
+                            Card.Type = cardType;
+                            doNotAddCard = await DisplayAlert("Please confirm:", $"\nCard Type: {Card.Type}\nAccount Number: {Card.AccountNumber}", "No", "Yes");
+                        }
+                        break;
+                    }
+                case "Other":
+                    {
+                        if (Card.AccountNumber == null || Card.AccountNumber == "")
+                        {
+                            await DisplayAlert("Add Incomplete", "Please include values for each field.", "Ok");
+                        }
+                        else
+                        {
+                            if (frontImage != null)
+                                Card.FrontImage = frontImage;
+                            if (backImage != null)
+                                Card.BackImage = backImage;
                             Card.Type = cardType;
                             doNotAddCard = await DisplayAlert("Please confirm:", $"\nCard Type: {Card.Type}\nAccount Number: {Card.AccountNumber}", "No", "Yes");
                         }
@@ -157,6 +196,30 @@ namespace cardkeeper.ViewModels
                     return;
                 frontImage = System.IO.File.ReadAllBytes(file.Path);
                 DisplayFrontImage = Converter.ByteToImage(frontImage);
+                File.Delete(file.Path);
+                file.Dispose();
+            }
+
+        }
+        public async void TakeBackPhoto()
+        {
+            if (CrossMedia.Current.IsCameraAvailable && CrossMedia.Current.IsTakePhotoSupported)
+            {
+                // Supply media options for saving our photo after it's taken.
+                var mediaOptions = new Plugin.Media.Abstractions.StoreCameraMediaOptions
+                {
+                    Directory = "Cards",
+                    Name = $"{DateTime.UtcNow}.jpg",
+                    PhotoSize = Plugin.Media.Abstractions.PhotoSize.Custom,
+                    CustomPhotoSize = 40,
+                    AllowCropping = true,
+                };
+
+                var file = await CrossMedia.Current.TakePhotoAsync(mediaOptions);
+                if (file == null)
+                    return;
+                backImage = System.IO.File.ReadAllBytes(file.Path);
+                DisplayBackImage = Converter.ByteToImage(backImage);
                 File.Delete(file.Path);
                 file.Dispose();
             }
